feat(web): implement CSV data import with column mapping (#611)\n\nExtract inline step components from ImportPage into dedicated reusable\ncomponents under components/import/:\n\n- StepIndicator: wizard progress nav with aria-current\n- FileDropZone: reusable drag-and-drop file upload with keyboard support\n- ColumnMapper: column-to-field mapping table with auto-detection\n- ImportPreview: validation summary, error table, duplicate detection\n- ImportProgress: accessible progress bar with role=progressbar\n- ImportComplete: import results summary with navigation\n\nThe CSV import feature includes:\n- Pure-TypeScript CSV parser with auto-delimiter detection (RFC 4180)\n- Intelligent column mapping with confidence scores\n- Date parsing (ISO, US, EU formats) and amount parsing to cents\n- Duplicate detection by date+amount+description matching\n- Multi-step wizard (upload, map, preview, import, complete)\n- Batch import with progress tracking\n- Full accessibility: ARIA, keyboard nav, screen reader announcements\n- Responsive design with dark mode, reduced motion, high contrast\n\nCo-authored-by: Claude <noreply@anthropic.com>
